requirements.txt
spotipy==2.23.0
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0
pytest==7.4.0
jupyter==1.0.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn==1.2.2